Moved additional plotting functions into new plotting class from sygma
