+++ sygma.py
@@ -277,397 +277,6 @@ class sygma( chem_evol, plotting ):
 ######################## Here start the analysis methods ######################################
 ###############################################################################################
 
-    def __plot_mass_multi(self,fig=1,specie=['C'],ylims=[],source='all',norm=False,label=[],shape=['-'],marker=['o'],color=['r'],markevery=20,fsize=[10,4.5],fontsize=14,rspace=0.6,bspace=0.15,labelsize=15,legend_fontsize=14):
-
-    	'''
-	Use the function plot_mass multiple times
-	Mass evolution (in Msun) of an element or isotope vs time.
-	
-
-        Parameters
-        ----------
-
-
-	yaxis : array
-             isotopes or element names, in the form 'C' or 'C-12'
-	source : string
-             Specifies if yields come from
-	     all sources ('all'), including
-	     AGB+SN1a, massive stars. Or from
-	     distinctive sources:
-	     only agb stars ('agb'), only
-	     SN1a ('SN1a'), or only massive stars
-             ('massive')
-        norm : boolean
-             If True, normalize to current total ISM mass
-	label : string
-	     figure label
-	marker : string
-	     figure marker
-	shape : string
-	     line style	
-	color : string
-	     color of line
-        fig : string,float
-	     to name the plot figure       
- 	       
-        Examples
-	----------
-
-	'''
-	#fig=plt.figure(fig)
-	#nplots=1#len(specie)
-	#f, ax_plots = plt.subplots(nplots, sharex=True, sharey=False)
-	limits=[]
-	ax=plt.gca()
-	props = dict(boxstyle='square', facecolor='w', alpha=1)
-	for k in range(len(specie)):
-		#if len(ylims)>0:
-			#ylims1=ylims[k]
-			#ax_plots[k].set_ylim(ylims[k][0],ylims[k][1])
-		if len(label)==0:
-			label1=specie[k]
-		else:
-			label1=label[k]
-		x,y=self.plot_mass(fig=fig,specie=specie[k],source=source,norm=norm,label=label1,shape=shape[k],marker=marker[k],color=color[k],markevery=20,multiplot=True)
-		#x=np.log10(np.array(x))
-		y=np.log10(np.array(y))
-		#ax_plots[k]
-		plt.plot(x,y,label=label1,linestyle=shape[k],marker=marker[k],color=color[k],markevery=markevery)
-		#ax_plots[k].set_ylim(min(y),max(y))
-		limits.append([min(y),max(y)])
-		#ax_plots[k].set_xlim(min(x),max(x))
-		#ax_plots[k].
-		plt.xscale('log')
-		#ax_plots[k].set_yscale('log')
-		plt.xlabel('Log-scaled age [yrs]')
-		#ax_plots[k].locator_params(axis = 'y', nbins = 2)
-		#if norm == False:
-		    #ax_plots[k].set_ylabel('Log-scaled yields [Msun]')
-		    #ax_plots[k].set_yscale('log')
-		#else:
-		#    ax_plots[k].set_ylabel('(IMF-weighted) fraction of ejecta')
-		#plt.legend()
-                self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize,lwtickboth=[3,1],lwtickmajor=[5,1])
-		#ax_plots[k].
-		#plt.legend().set_visible(False)
-		#x.legend(loc='center right', bbox_to_anchor=(1, 0.5),markerscale=0.8,fontsize=fontsize).set_visible(False)
-		#ax_plots[k]
-		#.text(0.90, 0.80, label1, transform=ax_plots[k].transAxes, fontsize=18,verticalalignment='top', bbox=props)
-		plt.xlim(self.history.dt,self.history.tend)
-	if norm == False:
-		fig=plt.gcf()
-		fig.text(0.002, 0.5, 'Log (Yields [Msun])', ha='center', va='center', rotation='vertical')
-	else:
-		fig=plt.gcf()
-		fig.text(0.01, 0.5, '(IMF-weighted) fraction of ejecta', ha='center', va='center', rotation='vertical')
-
-	#set lim here
-	#print limits
-	#for k in range(len(ax_plots)):
-	#	ax_plots[k].set_ylim(limits[k][0],limits[k][1])
-
-	#f.subplots_adjust(hspace=0.35)#0)
-	#plt.setp([a.get_xticklabels() for a in f.axes[:-1]], visible=False)
-	return 
-
-    def __plot_abu_distr(self,fig=0,t=-1,x_axis='A',solar_norm=True,marker1=2,linest=0,y_range=[],label='CHEM  module',fsize=[10,4.5],fontsize=14,rspace=0.6,bspace=0.15,labelsize=15,legend_fontsize=14):
-
-        '''
-	EXPERIMENTAL: DO NOT USE
-	Plots abundance distribution of stable isotopes
-	of certain time t (X/X_sol vs A).:
-
-        Parameters
-        ----------
-
-	t : float
-            default end of the simulation tend
-	x_axis :  string
-            only 'A' setting, mass number, possible
-        solar_norm : boolean
-            if true, normalize to solar value
-        linest : string
-            linestyle
-        makre1 : string
-            for marker
-        y_range : list
-            contains min and max
-
-        Examples
-        ----------
-	>>> s.plot_abu_distr()
-
-        '''
-
-
-        color=['r','k','b','g']
-        marker_type=['o','+','s','D']
-        line_style=['--','-','-.',':']
-
-
-        x=[]
-        y=[]
-        fig=plt.figure(fig,figsize=(10,8))
-        ax = fig.add_subplot(1,1,1)
-        if len(y_range)>0:
-            plt.ylim(y_range[0],y_range[1])
-        elem_array=[]
-        y_array=[]
-        iso_name='H-1'
-        ele='H'
-        i=0
-        colormap = plt.cm.prism#flag #gist_ncar
-        plt.gca().set_color_cycle([colormap(j) for j in np.linspace(0, 0.9, 100)])
-
-
-        if t==-1:
-            t=self.history.tend
-        idx=int(t/self.history.dt) -1
-        isotopes1=self.history.isotopes
-        yields_iso1=self.history.ism_iso_yield[idx]
-        isotopes=[]
-        yields_iso=[]
-
-        #Check for stable isotopes
-        from nugrid_set import is_stable
-        for k in range(len(isotopes1)):
-            if is_stable(isotopes1[k]):
-                isotopes.append(isotopes1[k])
-                yields_iso.append(yields_iso1[k])
-
-
-        iniabu=ry.iniabu('yield_tables/iniabu/iniab2.0E-02GN93.ppn')
-        x_ini_iso=iniabu.iso_abundance(isotopes)
-        #elements,x_ini=self.iso_abu_to_elem(self.isotopes,x_ini_iso)
-
-        for k in range(len(isotopes)):
-            elem_name=isotopes[k].split('-')[0]
-            if x_axis=='A':
-                # Xi / Xsun
-                yields_norm=(yields_iso[k]/sum(yields_iso))/x_ini_iso[isotopes.index(isotopes[k])]
-
-                #in the case of isotope of same element as before
-                if elem_name==ele:
-                    elem_array.append(isotopes[k].split('-')[1] )
-                    y_array.append(yields_norm )
-                    print 'new iso',isotopes[k]
-                    #i+=1
-                    if not k==(len(isotopes)-1):
-                        continue
-                print k
-                plt.plot(elem_array,y_array,marker=marker_type[marker1],markersize=8,linestyle=line_style[linest])
-                print elem_array
-                print y_array[0]
-                i+=1
-                if (i%2) ==0:
-                    high=6
-                else:
-                    high=4
-                print 'i',i
-                print 'high',high
-                y_pos=(max(y_array))
-                x_pos=  elem_array[y_array.index(max(y_array))]
-                ax.annotate(iso_name, xy=(x_pos, y_pos), xytext=(elem_array[0],high))
-
-                ele=elem_name
-                iso_name=isotopes[k]
-                y_array=[]
-                y_array.append(yields_norm)
-                elem_array=[]
-                elem_array.append(isotopes[k].split('-')[1] )
-                #x.append(line_1[1])
-                plt.plot([0,72],[1,1],linestyle='--')
-                plt.plot([0,72],[2,2],linestyle='--')
-                plt.plot([0,72],[0.5,0.5],linestyle='--')
-
-
-
-
-
-            #if x_axis=='Z':
-            #       x.append(line_1[0])
-
-        plt.xlabel("Mass number A")
-            #plt.plot(x,y,marker=marker_type[0],markersize=10,linestyle='None',label=self.label)
-        ax.set_yscale('log')
-        if len(y_range)>0:
-            plt.ylim(y_range[0],y_range[1])
-        plt.xlim(0,72)
-        simArtist = plt.Line2D((0,1),(0,0), mfc='none',marker=marker_type[marker1], linestyle='w')
-        ax=plt.gca()
-        plt.legend([simArtist],[label])
-        ax=plt.gca()
-        self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-
-    def __plot_mass_range_contributions_single(self,fig=7,specie='C',prodfac=False,rebin=1,time=-1,label='',shape='-',marker='o',color='r',markevery=20,extralabel=False,log=False,fsize=[10,4.5],fontsize=14,rspace=0.6,bspace=0.15,labelsize=15,legend_fontsize=14):
-
-        '''
-	Internal plotting function for function plot_mass_range_contributions
-        '''
-    
-
-        if len(label)==0:
-            label=specie
-
-
-        contribution=[]
-        #find starburst corresponding to time
-        if time>0:
-                age1=self.history.age
-                if time in age1:
-                        age_idx=age1.index(time)
-                else:
-                        age_idx=min(range(len(age1)), key=lambda i: abs(age1[i]-time))
-                        print 'Age not found, choose closest age',age1[age_idx]
-                mass_ranges=self.history.imf_mass_ranges[age_idx]
-                contribution=self.history.imf_mass_ranges_contribution[age_idx]
-                mtots=self.history.imf_mass_ranges_mtot[age_idx]
-        #take sum of all star bursts
-        else:
-
-                firstTime=True
-                for k in range(len(self.history.imf_mass_ranges)):
-                        mass_ranges1=self.history.imf_mass_ranges[k]
-                        if len(mass_ranges1)>0 and firstTime==True:
-                                mass_ranges=self.history.imf_mass_ranges[k]
-                                contribution=list(self.history.imf_mass_ranges_contribution[k])
-                                mtots=self.history.imf_mass_ranges_mtot[k]
-                                firstTime=False
-                        elif len(mass_ranges1)>0 and (not len(self.history.imf_mass_ranges[k]) == len(mass_ranges)):
-                                print 'Error: Different mass range intervalls used: cannot combine them'
-                                print 'Choose a specific staburst via time'
-                                return 0
-                        elif len(mass_ranges1)>0:
-                                mtots=np.array(mtots)+np.array(self.history.imf_mass_ranges_mtot[k])
-                                #carries all isotopes, hence an array: self.history.imf_mass_ranges_contribution[k][h]
-                                for h in range(len(self.history.imf_mass_ranges_contribution[k])):
-                                        contribution[h]= np.array(contribution[h])+np.array(self.history.imf_mass_ranges_contribution[k][h])
-
-
-        isotopes=self.history.isotopes
-        iso_idx_array=[]
-	x_iso_ism_ini=np.array(self.history.ism_iso_yield[0])/self.history.mgal
-        #get the element specific indices
-        for k in range(len(isotopes)):
-	    if not '-' in specie:
-	    	specie=specie+'-'
-            if specie in isotopes[k]:
-                iso_idx_array.append(isotopes.index(isotopes[k]))
-        #get the sum of yields for each isotope and mass range
-        y=[0]*len(contribution)
-        for k in range(len(contribution)):
-	    yields=0 #[0]*len(contribution[k])
-	    x_ini=0
-            for iso_idx in iso_idx_array:
-                yields+= np.array(contribution[k][iso_idx])
-		x_ini+=x_iso_ism_ini[iso_idx]
-	    if prodfac==True:
-		if x_ini ==0:
-			print 'Initial abundance not available!'
-			print 'Cannot plot production factor.'
-			return 0
-		y[k]=np.array(yields)/(mtots[k]*x_ini)
-	    else:
-	 	y[k]=yields
-		
-	#print mass_ranges
-	#print 'contribution'
-	#print y
-
-        #masses1=[]
-        #for m in range(len(mass_ranges)):
-        #    masses1.append( (mass_ranges[m][0]+mass_ranges[m][1])/2.)
-        #masses_idx= sorted(range(len(masses1)),key=lambda x:masses1[x])
-
-        #print 'len same: ',len(mass_ranges),len(masses)
-        #for idx in masses_idx:
-        #    masses.append(masses1[idx])
-        #    yields.append(yields1[idx])
-        #'''
-        #for histo
-        bin_bdys=[]
-	mean_val=[]
-        #bin_bdys.append(1)
-        bin_values=[]
-	#rebin the plot using the bin size rebin
-	#print mass_ranges
-	if rebin >0:
-		#print mass_ranges
-		#print y
-		#nprint '-------------------'
-		#_imf(self,mmin,mmax,inte,mass=0,iolevel=0)
-		mass=mass_ranges[0][0]	
-		mmax=mass_ranges[-1][1]
-		bin_bdys1=[]
-		while True:
-			bin_min=round(mass,5)
-			mass+=rebin
-			bin_max=round(mass,5)
-			if (mmax==0):
-				break
-			if bin_max>mmax:
-				bin_max=mmax
-				mass=mmax
-				mmax=0
-			bin_bdys1.append([bin_min,bin_max])
-			bin_values.append(0)
-			#print 'Bin :',bin_bdys1[-1]
-			for k in range(len(mass_ranges)):
-				if (mass_ranges[k][1]<=bin_min) or (mass_ranges[k][0]>=bin_max):
-					continue
-				#print 'interval ',mass_ranges[k]
-				#if mass range inside bin
-				elif (mass_ranges[k][0]>=bin_min) and (mass_ranges[k][1]<=bin_max):
-					bin_values[-1]+=y[k]
-					#print 'bin includes mass range',y[k]
-					continue
-				#if mass range includes bin:
-				elif (mass_ranges[k][0]<=bin_min) and (mass_ranges[k][1]>=bin_max):
-                                        #normalization to bin mass
-                                        h=y[k]/self._imf(mass_ranges[k][0],mass_ranges[k][1],inte=2)
-                                        y1=h*self._imf(bin_min,bin_max,inte=2)
-                                        bin_values[-1]+=y1
-                                        #print 'mass range inlcudes bin ',bin_min,bin_max,y1
-					
-				#if upper part of bin is not in mass range
-				elif (mass_ranges[k][1]<bin_max):
-					#normalization to bin mass
-					h=y[k]/self._imf(mass_ranges[k][0],mass_ranges[k][1],inte=2)
-					y1=h*self._imf(bin_min,mass_ranges[k][1],inte=2)
-					bin_values[-1]+=y1
-					#print 'add upper half from ',bin_min,mass_ranges[k][1],y1
-                                #if lower part of bin is not in mass range
-                                elif mass_ranges[k][0]>bin_min:					
-					#normalization to bin mass
-					h=y[k]/self._imf(mass_ranges[k][0],mass_ranges[k][1],inte=2)
-					y1=h*self._imf(mass_ranges[k][0],bin_max,inte=2)
-					#print 'add lower half from ',mass_ranges[k][0],bin_max,y1
-					bin_values[-1]+=y1
-			#if bin_values[-1]==0:
-			#	print 'Note that no values are found in bin range:',bin_min,'-',bin_max
-				#return 0
-			if mmax==bin_max:
-				break
-	#	print bin_bdys1
-#		print bin_values
-		mass_ranges=bin_bdys1
-		y=bin_values
-
-
-	for k in range(len(mass_ranges)):
-                #yields1.append(yields[k]) #
-		if k==0:
-                	bin_bdys.append(mass_ranges[k][0])
-                bin_bdys.append(mass_ranges[k][1])
-		mean_val.append( (mass_ranges[k][0] + mass_ranges[k][1])/2.)
-	#print 'test'		
-	#print yields
-	#print bin_bdys
-	#print mean_val
-
-	return mean_val,bin_bdys,y,color,label
-
     def save_data(self,header=[],data=[],filename='plot_data.txt'):
 	'''
 		Writes data into a text file. data entries
